Move argument_parsing.hpp to common/
- Add `common/` to include path for example exes.
