+++ examples/common/argument_parsing.hpp
